+++ src/main/php/InvocationThrow.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of bovigo\callmap.
  *
@@ -35,9 +36,8 @@ class InvocationThrow
      *
      * @return  \Exception
      */
-    public function exception()
+    public function exception(): \Exception
     {
         return $this->e;
     }
 }
-