+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### BC breaks
 
   * raised minimum required PHP version to 7.0.0
+  * introduced scalar type hints and strict type checking
 
 
 2.1.0 (2016-06-20)